[CHORE]: 작품 설명 화면 디자인 수정
