+++ ARtique/ARtique/Screen/Add/Cell/CVC/ArtworkExplainCVC.xib
@@ -11,17 +11,17 @@
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
         <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="ArtworkExplainCVC" id="gTV-IL-0wX" customClass="ArtworkExplainCVC" customModule="ARtique" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="300" height="527"/>
+            <rect key="frame" x="0.0" y="0.0" width="300" height="482"/>
             <autoresizingMask key="autoresizingMask"/>
             <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
-                <rect key="frame" x="0.0" y="0.0" width="300" height="527"/>
+                <rect key="frame" x="0.0" y="0.0" width="300" height="482"/>
                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                 <subviews>
                     <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="E17-rw-cau">
-                        <rect key="frame" x="0.0" y="0.0" width="300" height="527"/>
+                        <rect key="frame" x="0.0" y="0.0" width="300" height="482"/>
                         <subviews>
                             <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Dui-K3-Pga">
-                                <rect key="frame" x="0.0" y="0.0" width="300" height="527"/>
+                                <rect key="frame" x="0.0" y="0.0" width="300" height="482"/>
                                 <subviews>
                                     <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="Fvc-fi-4hD">
                                         <rect key="frame" x="0.0" y="0.0" width="300" height="300"/>
@@ -38,10 +38,10 @@
                                         <textInputTraits key="textInputTraits"/>
                                     </textField>
                                     <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" scrollEnabled="NO" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="4dv-2j-TT9">
-                                        <rect key="frame" x="0.0" y="372" width="300" height="155"/>
+                                        <rect key="frame" x="0.0" y="372" width="300" height="110"/>
                                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                                         <constraints>
-                                            <constraint firstAttribute="height" constant="155" id="qG9-aq-fE1"/>
+                                            <constraint firstAttribute="height" constant="110" id="qG9-aq-fE1"/>
                                         </constraints>
                                         <color key="textColor" systemColor="labelColor"/>
                                         <fontDescription key="fontDescription" type="system" pointSize="14"/>
@@ -64,11 +64,11 @@
                             </view>
                         </subviews>
                         <constraints>
-                            <constraint firstItem="Dui-K3-Pga" firstAttribute="bottom" secondItem="WlS-HG-lYs" secondAttribute="bottom" id="LL0-tG-4Zd"/>
-                            <constraint firstItem="Dui-K3-Pga" firstAttribute="trailing" secondItem="WlS-HG-lYs" secondAttribute="trailing" id="dhQ-wP-uCH"/>
-                            <constraint firstItem="Dui-K3-Pga" firstAttribute="top" secondItem="WlS-HG-lYs" secondAttribute="top" id="lcR-5u-itx"/>
+                            <constraint firstItem="Dui-K3-Pga" firstAttribute="top" secondItem="WlS-HG-lYs" secondAttribute="top" id="DNs-Rm-XsE"/>
+                            <constraint firstItem="Dui-K3-Pga" firstAttribute="leading" secondItem="WlS-HG-lYs" secondAttribute="leading" id="JLQ-W1-9GY"/>
+                            <constraint firstItem="Dui-K3-Pga" firstAttribute="bottom" secondItem="WlS-HG-lYs" secondAttribute="bottom" constant="482" id="aKo-B0-hAH"/>
                             <constraint firstItem="Dui-K3-Pga" firstAttribute="width" secondItem="LVW-71-Wn2" secondAttribute="width" id="rws-2w-YFH"/>
-                            <constraint firstItem="Dui-K3-Pga" firstAttribute="leading" secondItem="WlS-HG-lYs" secondAttribute="leading" id="yhO-vB-EBn"/>
+                            <constraint firstItem="Dui-K3-Pga" firstAttribute="trailing" secondItem="WlS-HG-lYs" secondAttribute="trailing" constant="300" id="sSe-Pr-2q6"/>
                         </constraints>
                         <viewLayoutGuide key="contentLayoutGuide" id="WlS-HG-lYs"/>
                         <viewLayoutGuide key="frameLayoutGuide" id="LVW-71-Wn2"/>
@@ -78,18 +78,18 @@
             <viewLayoutGuide key="safeArea" id="ZTg-uK-7eu"/>
             <constraints>
                 <constraint firstItem="E17-rw-cau" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" id="3gQ-b0-VBe"/>
-                <constraint firstAttribute="bottom" secondItem="E17-rw-cau" secondAttribute="bottom" id="c1O-be-HAb"/>
                 <constraint firstItem="E17-rw-cau" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" id="n2b-wj-Y3o"/>
+                <constraint firstAttribute="bottom" secondItem="E17-rw-cau" secondAttribute="bottom" id="pei-R1-fQf"/>
                 <constraint firstAttribute="trailing" secondItem="E17-rw-cau" secondAttribute="trailing" id="x0E-Fm-soU"/>
             </constraints>
-            <size key="customSize" width="232" height="417"/>
+            <size key="customSize" width="232" height="386"/>
             <connections>
                 <outlet property="contentTextView" destination="4dv-2j-TT9" id="5eo-If-TW1"/>
                 <outlet property="image" destination="Fvc-fi-4hD" id="2RN-qV-mmL"/>
                 <outlet property="scrollView" destination="E17-rw-cau" id="4js-VI-ZZX"/>
                 <outlet property="titleTextField" destination="bDq-NE-jEL" id="x3F-Fh-dMO"/>
             </connections>
-            <point key="canvasLocation" x="269.56521739130437" y="239.39732142857142"/>
+            <point key="canvasLocation" x="269.56521739130437" y="229.01785714285714"/>
         </collectionViewCell>
     </objects>
     <resources>
